// web/src/views/system/menu/utils/types.ts
interface FormItemProps {
  /** 菜单类型（0代表菜单、1代表iframe、2代表外链、3代表按钮）*/
  id?: number;
  menuType: number;
  higherMenuOptions: Record<string, unknown>[];
  parentId: number;
  title: string;
  name: string;
  path: string;
  component: string;
  rank: number;
  redirect: string;
  icon: string;
  extraIcon: string;
  enterTransition: string;
  leaveTransition: string;
  activePath: string;
  auths: string;
  frameSrc: string;
  frameLoading: number;
  keepAlive: number;
  hiddenTag: number;
  fixedTag: number;
  showLink: number;
  showParent: number;
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };